moved job specific code into its own branch. This branch now has just the Fluent Scheduler implementation with examples
