+++ CSLabs.Api/Jobs/JobRegistry.cs
@@ -8,11 +8,9 @@ namespace CSLabs.Api.Jobs
     {
         public JobRegistry(IServiceProvider provider)
         {
-            // Every 15 seconds check the status of each Lab VM within each Lab
-            // Every 3 minutes check to see if we have quorum (at least 50% of the ProxMox nodes are up)
             
             // Schedule new jobs here
-            Schedule(() => new NodeStatusJob(provider)).ToRunEvery(15).Seconds();
+            Schedule(() => new ExampleJob(provider)).ToRunEvery(1).Minutes();
 
         }
     }
